move ExpectedTableOwner to table package
