+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/segment/impl/table/ExpectedTableOwner.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.owner;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.table;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -23,6 +23,11 @@ import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.AbstractExpec
 
 import javax.xml.bind.annotation.XmlAttribute;
 
+/**
+ * Expected table owner.
+ * 
+ * @author zhangliang 
+ */
 @Getter
 @Setter
 public final class ExpectedTableOwner extends AbstractExpectedDelimiterSQLSegment {
